Update ja_jp.json [Need discussion] (Read description)
### Mod name:
Currently, it is "リフォームの匠" which I think should definitely be changed. 匠 is like an unofficial name for creepers in japanese, which also means like 'master of', so it could mean the name is a play on words BUT, it would still mean like "master of reforming" or "creeper of reforming (inaccurate grammar)". I suggest like "クリーパーオーバーホール" which literally is "Creeper Overhaul" (like a loan word) or "クリーパー改造" which means more like "Creeper Modification" (since overhaul isn't used that commonly in japanese)

### Creeper names:
All the names are more playful and are like mostly combining the words together like bamboo and creeper to something like bamboreeper etc. But, the badlands creeper translation roughly translates to teracopper which is combining teracotta and creeper, not the biome badlands (which in japanese is "荒野" (pronounced 'arano' or 'kouya')).
The mushroom creeper is also called 'Boom-room'. 
Overall there are many inconsistency with the names with hill, dark oak, spruce, beach creeper names just being '[theme name] creeper'. I think it should all be like '[theme name] creeper' to be more consistent.